+++ src/main/java/com/slam/concertreservation/domain/queue/model/Token.java
@@ -1,6 +1,6 @@
 package com.slam.concertreservation.domain.queue.model;
 
-import com.slam.concertreservation.exceptions.BusinessRuleViolationException;
+import com.slam.concertreservation.common.exceptions.BusinessRuleViolationException;
 import java.time.LocalDateTime;
 import java.util.Objects;
 import lombok.Getter;
