+++ src/main/java/com/slam/concertreservation/domain/reservation/model/Reservation.java
@@ -1,6 +1,6 @@
 package com.slam.concertreservation.domain.reservation.model;
 
-import com.slam.concertreservation.exceptions.BusinessRuleViolationException;
+import com.slam.concertreservation.common.exceptions.BusinessRuleViolationException;
 import java.time.LocalDateTime;
 import java.util.UUID;
 import lombok.Getter;
