+++ src/main/java/com/slam/concertreservation/domain/reservation/service/ReservationRollbackService.java
@@ -4,7 +4,7 @@ import com.slam.concertreservation.domain.reservation.event.ReservationConfirmat
 import com.slam.concertreservation.domain.reservation.event.ReservationExpirationRollbackEvent;
 import com.slam.concertreservation.domain.reservation.model.Reservation;
 import com.slam.concertreservation.domain.reservation.repository.ReservationRepository;
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.context.ApplicationEventPublisher;
