+++ src/main/java/com/slam/concertreservation/infrastructure/outbox/OutboxService.java
@@ -1,6 +1,6 @@
 package com.slam.concertreservation.infrastructure.outbox;
 
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
