+++ src/main/java/com/slam/concertreservation/infrastructure/persistence/redis/impl/TokenRepositoryRedisImpl.java
@@ -3,8 +3,7 @@ package com.slam.concertreservation.infrastructure.persistence.redis.impl;
 import com.slam.concertreservation.domain.queue.model.Token;
 import com.slam.concertreservation.domain.queue.model.TokenStatus;
 import com.slam.concertreservation.domain.queue.repository.TokenRepository;
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
-import jakarta.annotation.Resource;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import java.time.LocalDateTime;
 import java.time.ZoneId;
 import java.util.ArrayList;
@@ -12,7 +11,6 @@ import java.util.List;
 import java.util.Optional;
 import java.util.Set;
 import java.util.UUID;
-import lombok.RequiredArgsConstructor;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.data.redis.core.Cursor;
 import org.springframework.data.redis.core.HashOperations;
