+++ src/main/java/com/slam/concertreservation/infrastructure/persistence/redis/locking/RedissonDistributedLockAop.java
@@ -1,6 +1,6 @@
 package com.slam.concertreservation.infrastructure.persistence.redis.locking;
 
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.aspectj.lang.ProceedingJoinPoint;
