+++ src/test/java/com/slam/concertreservation/application/ConcurrentConcertReservationIntegrationTest.java
@@ -11,8 +11,8 @@ import com.slam.concertreservation.domain.concert.service.ConcertService;
 import com.slam.concertreservation.domain.point.model.UserPointBalance;
 import com.slam.concertreservation.domain.reservation.repository.ReservationRepository;
 import com.slam.concertreservation.domain.user.model.User;
-import com.slam.concertreservation.exceptions.BusinessRuleViolationException;
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.BusinessRuleViolationException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import java.util.concurrent.atomic.AtomicInteger;
 import org.junit.jupiter.api.*;
 import org.slf4j.Logger;
