+++ src/test/java/com/slam/concertreservation/domain/concert/model/SeatUnitTest.java
@@ -3,7 +3,7 @@ package com.slam.concertreservation.domain.concert.model;
 import static org.assertj.core.api.Assertions.assertThatNoException;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 
-import com.slam.concertreservation.exceptions.BusinessRuleViolationException;
+import com.slam.concertreservation.common.exceptions.BusinessRuleViolationException;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
