+++ src/test/java/com/slam/concertreservation/domain/concert/service/ConcertServiceUnitTest.java
@@ -7,8 +7,8 @@ import com.slam.concertreservation.domain.concert.model.SeatStatus;
 import com.slam.concertreservation.domain.concert.repository.ConcertRepository;
 import com.slam.concertreservation.domain.concert.repository.ConcertScheduleRepository;
 import com.slam.concertreservation.domain.concert.repository.SeatRepository;
-import com.slam.concertreservation.exceptions.BusinessRuleViolationException;
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.BusinessRuleViolationException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import java.time.LocalDateTime;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
