+++ src/test/java/com/slam/concertreservation/domain/point/service/PointServiceUnitTest.java
@@ -6,7 +6,7 @@ import com.slam.concertreservation.domain.point.model.PointTransactionType;
 import com.slam.concertreservation.domain.point.model.UserPointBalance;
 import com.slam.concertreservation.domain.point.repository.PointHistoryRepository;
 import com.slam.concertreservation.domain.point.repository.UserPointBalanceRepository;
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
