+++ src/test/java/com/slam/concertreservation/domain/reservation/service/ReservationServiceUnitTest.java
@@ -3,8 +3,8 @@ package com.slam.concertreservation.domain.reservation.service;
 import com.slam.concertreservation.domain.reservation.model.Reservation;
 import com.slam.concertreservation.domain.reservation.model.ReservationStatus;
 import com.slam.concertreservation.domain.reservation.repository.ReservationRepository;
-import com.slam.concertreservation.exceptions.BusinessRuleViolationException;
-import com.slam.concertreservation.exceptions.UnavailableRequestException;
+import com.slam.concertreservation.common.exceptions.BusinessRuleViolationException;
+import com.slam.concertreservation.common.exceptions.UnavailableRequestException;
 import java.time.LocalDateTime;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
